+++ packages/client-ledger/RELEASES.md
@@ -117,3 +117,13 @@ Based on:
 - [typescript v1.0.0-alpha.11] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.11] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.11 - packages/client-ledger
+
+## 2024-07-31 00:13:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.349.0 (2.382.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.12] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.12] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.12 - packages/client-ledger